requirements-local.txt
anyio==4.12.0
av==15.1.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.2
deep-translator==1.11.4
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai-whisper==20250625
packaging==25.0
pillow==11.3.0
protobuf==6.33.2
pycparser==2.23
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
scipy==1.13.1
shellingham==1.5.4
sounddevice==0.5.3
soupsieve==2.8
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.6.2
